requirements.txt
FLask
gunicorn
sklearn
pandas
numpy
pyYAML
-e .
